Refactoring of backend code from one big app.js file into smaller route files. Moved backend code into a src folder.
